# PERF: BooleanArray._from_sequence_of_strings